[fem] Move constitutive model test utils out of dev
