+++ multibody/fixed_fem/dev/BUILD.bazel
@@ -252,38 +252,6 @@ drake_cc_googletest(
     ],
 )
 
-drake_cc_library(
-    name = "constitutive_model_test_utilities",
-    testonly = 1,
-    srcs = ["test/constitutive_model_test_utilities.cc"],
-    hdrs = ["test/constitutive_model_test_utilities.h"],
-    deps = [
-        ":corotated_model",
-        ":linear_constitutive_model",
-        "//common:autodiff",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/test_utilities:expect_throws_message",
-        "//math:gradient",
-        "//multibody/fem:constitutive_model",
-    ],
-)
-
-drake_cc_googletest(
-    name = "corotated_model_test",
-    deps = [
-        ":constitutive_model_test_utilities",
-        ":corotated_model",
-    ],
-)
-
-drake_cc_googletest(
-    name = "linear_constitutive_model_test",
-    deps = [
-        ":constitutive_model_test_utilities",
-        ":linear_constitutive_model",
-    ],
-)
-
 drake_cc_googletest(
     name = "mesh_utilities_test",
     deps = [
